third_party/blink/web_tests/virtual/xslt-disabled/fast/xsl/xslt/xslt-in-subframe-quirk-expected.txt

CONSOLE WARNING: XSLTProcessor and XSLT Processing Instructions have been removed in this browser. See https://chromestatus.com/feature/4709671889534976.
Make sure that XSLT works in a subframe. 

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
